"FIx cassandra seeding so it has two tables, items and reviews"
