+++ database/cassandra/cassitemseedscript.cql
@@ -1,0 +1,3 @@
+USE mykeaitems;
+
+COPY items (id, title, colors, sizes, originalPrice, salePrice, description, liked, inStock) FROM '/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/csvsfiles/cassItemData.csv' with header=true and delimiter =';';
