+++ import_mykeaitems_itemreviews.err.20201022_160003
@@ -1,0 +1,68 @@
+const fs = require('fs');
+const _ = require('underscore');
+let documentNum = 10**7;
+let start = new Date();
+let samePrice = true;
+const writeReviews = fs.createWriteStream('/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/csvsfiles/cassReviewData.csv');
+writeReviews.write(`id,easeOfAssembly,valueForMoney,productQuality,appearance,worksAsExpected,overallRating,createdAt,iRecommendThisProduct,header,bodyn`, 'utf8');
+    let ok = true;
+      i--;
+      id++;
+      let easeOfAssembly = 0;
+      let valueForMoney = 0;
+      let productQuality = 0;
+      let appearance = 0;
+      let worksAsExpected = 0;
+      let overallRating = 0;
+      let createdAt = 0;
+      let iRecommendThisProduct = false;
+      let header = '';
+      let body = '';
+      let data = [];
+        easeOfAssembly = Math.floor(Math.random() * 5 + 1);
+        valueForMoney = Math.floor(Math.random() * 5 + 1);
+        productQuality = Math.floor(Math.random() * 5 + 1);
+        appearance = Math.floor(Math.random() * 5 + 1);
+        worksAsExpected = Math.floor(Math.random() * 5 + 1);
+        overallRating = Math.floor((easeOfAssembly + valueForMoney + productQuality + appearance + worksAsExpected) / 5);
+        createdAt = randomDate(new Date(2017, 0, 1), new Date());
+        iRecommendThisProduct = Math.random() > .5 ? true : false;
+        header = faker.lorem.words();
+        body = faker.lorem.words();
+      data = data.join('');
+        ok = writeReviews.write(data, 'utf-8');
+        writeReviews.once('drain', write);
+  write();
+  writeReviews.end();
+  console.log(`Reviews pipe closed! This seed script took ${new Date() - start} milliseconds to make ${documentNum} documents.`);
+});
+  return new Date(start.getTime() + Math.random() * (end.getTime() - start.getTime())).toDateString();
+
+
+
+
+
+
+
+
+
+      for (let j = 0; j < 2; j++) {
+        data.push(`${id};${easeOfAssembly};${valueForMoney};${productQuality};${appearance};${worksAsExpected};${overallRating};${createdAt};${iRecommendThisProduct};${header};${body}n`);
+const generateFakeReviewDataScript = (id, i, callback) => {
+  let write = () => {
+    do{
+      // taken from below for iteration count: Math.floor(Math.random() * 26)
+      }
+      if (i === 0) {
+        writeReviews.write(data, 'utf-8', callback)
+      } else {
+      }
+    } while (i > 0 && ok) {
+      if (i > 0) {
+      }
+    }
+  }
+}
+generateFakeReviewDataScript(0, documentNum, () => {
+function randomDate(start, end) {
+}
