+++ package.json
@@ -17,22 +17,25 @@
     "psqlgenDataSizeIds": "node ./database/postgreSQL/PSQLDataGenScripts/SizesIDDataGenerationScript.js",
     "psqlcreatdb": "psql postgres -U postgres < ./database/postgreSQL/schema.sql",
     "psqlseedpg": "psql postgres -U postgres -d product_options < ./database/postgreSQL/seedpg.sql",
-    "cdbdatagen": "node ./database/CouchDB/CDBDataGenScripts/itemDataGenScript.js",
-    "cdbdatabasegen": "node ./database/CouchDB/dbcreatescript.js",
-    "cdbseed": "node ./database/CouchDB/seedDB.js",
-    "cassGenData": "node ./database/cassandra/datagenscript.js",
+    "cassItemGenData": "node ./database/cassandra/itemdatagenscript.js",
+    "cassReviewGenData": "node ./database/cassandra/reviewdatagenscript.js",
     "cassDBgenData": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/cassdbgen.cql\"",
     "casstablegenData": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/casstablegen.cql\"",
-    "casstableseeding": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/cassseedscript.cql\""
+    "cassreviewstableseeding": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/cassreviewsseedscript.cql\"",
+    "cassitemtableseeding": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/cassitemseedscript.cql\"",
+    "cassReviewstableseeding": "cqlsh --file=\"/Users/susannah/Desktop/Programming/Bootcampfiles/Projects/SDC/ProductOptions/database/cassandra/cassreviewsseedscript.cql\"",
+    "cassServerStart": "nodemon ./server/cassserver.js"
   },
-  "author": "August Dolan",
+  "author": "Client:August Dolan, Backend:Susannah Marcus",
   "license": "ISC",
   "dependencies": {
     "axios": "^0.20.0",
+    "cassandra-driver": "^4.6.0",
     "express": "^4.17.1",
     "faker": "^5.1.0",
     "moment": "^2.29.0",
     "mongoose": "^5.10.6",
+    "nodemon": "^2.0.6",
     "pg": "^8.4.1",
     "pg-hstore": "^2.3.3",
     "react": "^16.13.1",
